Files for ai store/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
replicate==1.0.4
httpx==0.28.1
aiofiles==24.1.0
python-multipart==0.0.20
pydantic==2.10.6
python-dotenv==1.0.1
